requirements.txt
Flask
Flask-Cors
fpdf2
google-generativeai
gunicorn
mysql-connector-python
requests